+++ pm-spa/src/assets/scss/layout/_all.scss
@@ -9,3 +9,5 @@
 // modals
 @import './modals/_modal-right';
 @import './modals/_modal-center';
+
+@import './modals/_task-modal';
