+++ pm-spa/src/assets/scss/layout/modals/_modal-right.scss
@@ -10,102 +10,14 @@
     background: var(--color-black-light);
     box-shadow: -5px 0px 10px 5px rgba(0, 0, 0, 0.25);
 
-    &__header {
-        @include flex-space-between;
-
-        padding-bottom: 1.5rem;
-        border-bottom: var(--border-black);
-    }
-
-    &__complete {
-        @include flex-center-align;
-    }
-
     &__close {
+        position: absolute;
+        top: 2rem;
+        right: 2rem;
+
         font-size: 2rem;
         color: var(--color-grey);
         cursor: pointer;
-    }
-
-    &__body {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        grid-template-rows: 1fr 1fr 5fr 2fr;
-        grid-template-areas:
-            'title title'
-            'assigned-to due-date'
-            'description description'
-            'comments comments';
-
-        height: calc(100% - 20rem);
-    }
-
-    &__title {
-        position: relative;
-
-        @include flex-align-start;
-
-        grid-area: title;
-
-        border-bottom: var(--border-black);
-
-        textarea {
-            width: 100%;
-            overflow: hidden;
-
-            background: transparent;
-            border: none;
-            border: 1px solid transparent;
-            resize: none;
-
-            font-family: 'Montserrat';
-            font-size: 2rem;
-            color: var(--color-grey);
-            border-radius: 3px;
-            padding: 5px 7.5px;
-            margin: 1rem 0;
-            transition: var(--transition-2);
-
-            &:hover {
-                border: var(--border-black);
-            }
-
-            &:focus {
-                outline: none;
-                border: var(--border-black);
-            }
-        }
-    }
-
-    &__assigned-to {
-        @include flex-align-start;
-
-        grid-area: assigned-to;
-
-        font-size: 2rem;
-        border-bottom: var(--border-black);
-    }
-
-    &__due-date {
-        @include flex-align-start;
-
-        grid-area: due-date;
-
-        border-bottom: var(--border-black);
-    }
-
-    &__description {
-        grid-area: description;
-        border-bottom: var(--border-black);
-    }
-
-    &__comments {
-        grid-area: comments;
-        border-bottom: var(--border-black);
-    }
-
-    &__footer {
-        height: 15rem;
     }
 }
